third_party/blink/web_tests/http/tests/appcache/local-content-expected.txt

CONSOLE WARNING: Application Cache API manifest selection is deprecated and will be removed in M82, around April 2020. See https://www.chromestatus.com/features/6192449487634432 for more details.
CONSOLE WARNING: Application Cache API manifest selection is deprecated and will be removed in M82, around April 2020. See https://www.chromestatus.com/features/6192449487634432 for more details.
CONSOLE ERROR: line 28: Not allowed to load local resource: stdio.h
Test that documents loaded from application cache don't get access to local resources accidentally.

Should say SUCCESS:

SUCCESS
